Isolate the live run-duration clock into a self-contained leaf component
The live "Runtime: X.XX s" display was driven by a 50ms setInterval in
useRunLifecycle that called setRunDurationMs (App-level state) 20 times per
second for the whole run. Each update re-rendered the ENTIRE app, including
the full conversation. On a large session a single render can take ~750ms, so
the renders ran back-to-back and pegged the main thread for the entire run —
freezing the UI. It is worst during a no-stream wait (e.g. a custom node's
own internal LLM call), where the timer is the only thing forcing re-renders.

Fix: remove the App-level ticker and drive the live display from a new
self-contained <LiveRunClock> leaf that ticks its OWN local state (200ms), so
only that tiny node re-renders — never the app or the conversation. The final
duration is still set once when the run finishes, so behaviour is unchanged.
